Simplify `binary` module, remove unused function from `compat`
