--- esp-ieee802154/src/binary/c_types.rs
@@ -1,21 +1,0 @@
-#![allow(
-    non_camel_case_types,
-    non_snake_case,
-    non_upper_case_globals,
-    dead_code
-)]
-
-pub type c_uint = u32;
-pub type c_int = i32;
-pub type c_ulonglong = u32;
-pub type c_longlong = i32;
-pub type c_uchar = u8;
-
-pub type c_short = i16;
-pub type c_ushort = u16;
-pub type c_schar = i8;
-pub type c_char = u8;
-pub type c_long = i32;
-pub type c_ulong = u32;
-
-pub enum c_void {}
